+++ docs/conf.py
@@ -69,7 +69,7 @@ copyright = f'{year}, {author}'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the built documents.
-release = '0.8.7'  # The full version, including alpha/beta/rc tags.
+release = '0.8.8'  # The full version, including alpha/beta/rc tags.
 version = release.rsplit('.', maxsplit=1)[0]